Cleanup PSK usage.
